requirements.txt
python-telegram-bot==20.5
requests
PyPDF2
python-docx
